Split RemoteSetup.kt and ApptainerImage.kt
