+++ script-server/src/main/kotlin/org/geobon/hpc/HPCConnection.kt
@@ -3,6 +3,8 @@ package org.geobon.hpc
 import kotlinx.coroutines.*
 import org.geobon.pipeline.outputRoot
 import org.geobon.script.ScriptType
+import org.geobon.server.RemoteSetup
+import org.geobon.server.RemoteSetupState
 import org.geobon.server.ServerContext.Companion.scriptStubsRoot
 import org.geobon.server.ServerContext.Companion.scriptsRoot
 import org.geobon.server.plugins.Containers
