+++ script-server/src/test/kotlin/org/geobon/hpc/HPCStepTest.kt
@@ -8,6 +8,7 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.runTest
 import org.geobon.pipeline.*
 import org.geobon.script.ScriptType
+import org.geobon.server.RemoteSetupState
 import org.geobon.server.ServerContext
 import org.geobon.server.ServerContext.Companion.scriptsRoot
 import org.geobon.server.plugins.Containers
